+++ API/Api.BLL/Entities/WarehouseParcel.cs
@@ -5,7 +5,7 @@ namespace Api.BLL.Entities
     public class WarehouseParcel
     {
         public int WarehouseId { get; set; }
-        public uint ParcelId { get; set; }
+        public long ParcelId { get; set; }
 
         public Warehouse Warehouse { get; set; }
         public Parcel Parcel { get; set; }
